DEVAS-BM-helper: rename SentimentMenu to FictionBodyToolbar
