+++ assembl/static2/css/components/commentHelperButton.scss
@@ -1,0 +1,5 @@
+.comment-helper {
+  background-color: transparent;
+  border: none;
+  padding: 0;
+}
