+++ assembl/static2/tests/unit/components/debate/brightMirror/fictionBodyToolbar.spec.jsx
@@ -4,24 +4,24 @@ import initStoryshots from '@storybook/addon-storyshots';
 import { configure, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-import SentimentMenu from '../../../../../js/app/components/debate/brightMirror/sentimentMenu';
+import FictionBodyToolbar from '../../../../../js/app/components/debate/brightMirror/fictionBodyToolbar';
 
 // Import existing storybook data
-import { props } from '../../../../../js/app/stories/components/debate/brightMirror/sentimentMenu.stories'; // eslint-disable-line max-len
+import { props } from '../../../../../js/app/stories/components/debate/brightMirror/fictionBodyToolbar.stories'; // eslint-disable-line max-len
 
 // Separate the snapshots in directories next to each component
 // Name should match with the story name
 initStoryshots({
-  storyKindRegex: /^SentimentMenu$/
+  storyKindRegex: /^FictionBodyToolbar$/
 });
 
 configure({ adapter: new Adapter() });
 
-describe('<SentimentMenu /> - with shallow', () => {
+describe('<FictionBodyToolbar /> - with shallow', () => {
   let wrapper;
 
   beforeEach(() => {
-    wrapper = mount(<SentimentMenu {...props} />);
+    wrapper = mount(<FictionBodyToolbar {...props} />);
   });
 
   it('should render 4 inactive icons if no mysentiment', () => {
